chore: 유틸 제거하고 models에 옮김 추후에 Product, Cart, Discount 모델 세개로 나누면 좋을 듯
