+++ src/refactoring/models/cart.ts
@@ -1,31 +1,31 @@
 import { CartItem, Coupon, Product } from "../../types";
 
-// utils
-import { calculateTotalWithDiscount, caclulateTotalDiscount } from "../utils";
+/**
+ * Product
+ */
+export const isOutOfStock = (remainingStock: number) => remainingStock <= 0;
 
-export const calculateCartTotal = (
-  cart: CartItem[],
-  selectedCoupon: Coupon | null
-) => {
-  const { totalBeforeDiscount, totalAfterDiscount } =
-    calculateTotalWithDiscount(cart);
-  const { totalAfterCouponDiscount, totalDiscount } = caclulateTotalDiscount(
-    totalBeforeDiscount,
-    totalAfterDiscount,
-    selectedCoupon
-  );
-
-  return {
-    totalBeforeDiscount: Math.round(totalBeforeDiscount),
-    totalAfterDiscount: Math.round(totalAfterCouponDiscount),
-    totalDiscount: Math.round(totalDiscount),
-  };
+export const getRemainingStock = (cart: CartItem[], product: Product) => {
+  const cartItem = getCartItem(cart, product);
+  return product.stock - (cartItem?.quantity || 0);
 };
 
+export const getCartItem = (cart: CartItem[], product: Product) => {
+  return cart.find((item) => item.product.id === product.id);
+};
+
+/**
+ * Cart
+ */
 export const findCartItemByProductId = (
   prevCart: CartItem[],
   productId: string
 ) => prevCart.find((item) => item.product.id === productId);
+
+export const filteredCartItemByProductId = (
+  prevCart: CartItem[],
+  productId: string
+) => prevCart.filter((item) => item.product.id !== productId);
 
 export const increaseCartItemQuantity = (
   prevCart: CartItem[],
@@ -63,4 +63,93 @@ export const updateCartItemQuantity = (
       return item;
     })
     .filter((item): item is CartItem => item !== null);
+};
+
+export const calculateCartItemTotal = (item: CartItem) => {
+  const discount = getMaxApplicableDiscount(item);
+  return item.product.price * item.quantity * (1 - discount);
+};
+
+export const calculateCartTotal = (
+  cart: CartItem[],
+  selectedCoupon: Coupon | null
+) => {
+  const { totalBeforeDiscount, totalAfterDiscount } =
+    calculateTotalWithDiscount(cart);
+  const { totalAfterCouponDiscount, totalDiscount } = caclulateTotalDiscount(
+    totalBeforeDiscount,
+    totalAfterDiscount,
+    selectedCoupon
+  );
+
+  return {
+    totalBeforeDiscount: Math.round(totalBeforeDiscount),
+    totalAfterDiscount: Math.round(totalAfterCouponDiscount),
+    totalDiscount: Math.round(totalDiscount),
+  };
+};
+
+/**
+ * Discount
+ */
+export const getMaxDiscount = (
+  discounts: { quantity: number; rate: number }[]
+) => {
+  return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
+};
+
+export const getAppliedDiscount = (item: CartItem) => {
+  const { discounts } = item.product;
+  const { quantity } = item;
+  let appliedDiscount = 0;
+  for (const discount of discounts) {
+    if (quantity >= discount.quantity) {
+      appliedDiscount = Math.max(appliedDiscount, discount.rate);
+    }
+  }
+  return appliedDiscount;
+};
+
+export const getMaxApplicableDiscount = (item: CartItem) =>
+  item.product.discounts.reduce(
+    (maxDiscount, d) =>
+      item.quantity >= d.quantity && d.rate > maxDiscount
+        ? d.rate
+        : maxDiscount,
+    0
+  );
+
+export const calculateTotalWithDiscount = (cart: CartItem[]) => {
+  const totalBeforeDiscount = cart.reduce(
+    (total, item) => total + item.product.price * item.quantity,
+    0
+  );
+  const totalAfterDiscount = cart.reduce(
+    (total, item) => total + calculateCartItemTotal(item),
+    0
+  );
+
+  return {
+    totalBeforeDiscount,
+    totalAfterDiscount,
+  };
+};
+
+export const caclulateTotalDiscount = (
+  totalBeforeDiscount: number,
+  totalAfterDiscount: number,
+  selectedCoupon: Coupon | null
+) => {
+  const totalAfterCouponDiscount = selectedCoupon
+    ? selectedCoupon.discountType === "amount"
+      ? Math.max(0, totalAfterDiscount - selectedCoupon.discountValue)
+      : totalAfterDiscount * (1 - selectedCoupon.discountValue / 100)
+    : totalAfterDiscount;
+
+  const totalDiscount = totalBeforeDiscount - totalAfterCouponDiscount;
+
+  return {
+    totalAfterCouponDiscount,
+    totalDiscount,
+  };
 };
