--- src/refactoring/utils/index.ts
@@ -1,4 +1,0 @@
-export * from "./cart";
-export * from "./discount";
-export * from "./product";
-export * from "./stock";
